feat: WIP Inicio de implementação da classe para manipular arquivos.

+++ srcs/Server/HttpRequest.cpp
@@ -61,7 +61,7 @@ std::string         HttpRequest::GetBody(void) const {
     return this->_body;
 }
 
-size_t              HttpRequest::GetBodySize(void) const {
+int              HttpRequest::GetBodySize(void) const {
     return this->_bodySize;
 }
